fixed some bugs for playershipssprite; conversion from per second to per frame seems way off though
